+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### v1.3.0 - 2024-08-29
+* Add support for Natality Messaging
+* Add configuration to enable or disable Natality Messaging
+* Update BFDR to v1.0.0-preview.7 and VRDR to v4.3.0
+* Use message header validation from vrdr-dotnet library
+* Add POST and GET route handling for vital record type and IG version number, ex. MA/Bundle/BFDR/v2.0
+  * Validate the record aligns with the url vital record type on POST
+  * Return record submissions based on record type on GET
+
 ### v1.2.2 - 2024-08-27
 * Update library v4.2.2